services/astrologer_service/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0,<4.0
djangorestframework-simplejwt>=5.3.0,<6.0
psycopg2-binary>=2.9.6,<3.0
